Fictional pandemic data

An entirely artificially constructed data set and context designed for classroom discussion and analysis.

Should raise interesting scientific questions on how the results should, or should not, be interpreted. It should also raise questions on what might be done next.

A "pandemic" context is given in the details since the data were created during the first week of March, 2020.

Instructors might choose to invent their own context.

Format

A data frame with 100 rows and 4 variables

City

City for which the outcome data were recorded.

A

The percent survival rate for infected persons given medical treatment "A".

B

The percent survival rate for infected persons given medical treatment "B".

C

The percent survival rate for infected persons given medical treatment "C".

Details

One fictional narrative for this data set is as follows.

A virulent virus has led to a world wide pandemic and that the case fatality rate (proportion of those infected who die) is huge (say 6

Suppose that through a concerted and collaborative effort of health scientists worldwide, three different treatments have been developed for this group. All three treatments have been used at one time or another on numerous patients in this group from 100 different cities worldwide. The data are observational, in that they were simply collected and the treatment given noted. No information is available on why one treatment or another was given in any particular instance.

The recovery rates (as a percent) for the patients treated by each of the three treatments were simply recorded for each of the hundred cities and are available for analysis as the data frame pandemic.

Some obvious questions of interest are the comparisons of treatments. For example, is treatment A better than B? Than C? Is B better than C?

See also